Conflicts:
	drake/multibody/rigid_body_tree.cc
	drake/multibody/test/rigid_body_tree/rigid_body_collision_clique_test.cc

+++ drake/common/trajectories/test/piecewise_function_test.cc
@@ -18,7 +18,7 @@ class PiecewiseFunctionTester : public PiecewiseFunction {
 void TestPiecewiseFunctionTimeRelatedGetters(
     const PiecewiseFunctionTester& traj,
     const std::vector<double>& time) {
-  EXPECT_EQ(traj.getNumberOfSegments(), time.size() - 1);
+  EXPECT_EQ(traj.getNumberOfSegments(), static_cast<int>(time.size()) - 1);
 
   EXPECT_EQ(traj.getStartTime(), time.front());
   EXPECT_EQ(traj.getEndTime(), time.back());
